Top Level Makefile: build doxygen for libmetal and open-amp
Build the doxygen content from libmetal and open-amp submodules.

At this stage doxygen content is not styled as the Sphinx docs and is not
included in the Sphinx data model nor in PDF output.

Override the the html target to depend on the doxygen build and to
copy the result into the Sphinx html output dir.

Provide an html-fast target to bypass doxygen dependencies.

Readthedocs wants control over the Sphinx build process.
(unless you want to override the whole build & version process).

So define targets that can be used before (doxygen) and after
(doxygen_copy) the main Readthedocs build process.

Readthedocs will use its own BUILDDIR so we need to allow it to be
overridden.

Signed-off-by: Bill Mills <bill.mills@linaro.org>