Add a reference section to the sidebar
Based on what Zephyr does, this adds hard-coded links to the sidebar.
The sidebar is expanded in the new _templates/layout.html and defined by
the reference_links list is conf.py

Also adds page boiler plate for release tag vs latest w/ hints about how to switch.
(Readthedocs adds its own boilerplate for PRs so we don't need to.)

The reference URLs work on the desktop and for ReadTheDocs for:
* latest
* v* tags like v2023.04
* PR#s like 17 etc

Disable vcs_link as it is useless in the current config.

Notes of the reference URLs:

ReadTheDocs undefines canonical_url after our conf.py runs and sets
html_base instead.  So what we do does not matter but for the desktop
try to match what they will do.

Also html_base gets set to <server>/en/latest/ for PRs so it is pretty
useless. Luckily the READTHEDOCS_CANONICAL_URL env var is set correctly.
So our url_base var in conf.py will be correct.

Therefore, we expand the reference URLs in conf.py and not in
the layout template.

Also RTDS html_base ends in / so assume the same for our url_base and
latest_url.  If the user defines something that does not on the make
command line, fix it up.

Signed-off-by: Bill Mills <bill.mills@linaro.org>