.readthedocs.yaml: define pre_build and post_build jobs for doxygen
Doxygen should be built before the main build and the doxygen html output
should be copied into the main html output dir after the main build.

Before we had a .readthedocs.yaml file, the read the docs build process
fills in resonable defaults.  Now that we have that file we need to supply
these defaults ourselves.  So we init the submoudles, install the
requirements.txt, and ask for all formats.

Because are using the cmake for the doxygen build we need cmake and
libmetal cmake requires the specified libraries.

Signed-off-by: Bill Mills <bill.mills@linaro.org>