Fix version menu on ReadTheDocs
When we added the .readthedocs.yaml the version menu pop up broke.
It seams with the yaml file we are using the later versions of things
and right now there is a bug in RTDS with regard to jQuery not being
defined.

The suggested workaround is to add the sphinxcontrib.jquery extension.
Also add it to the requirements.txt so the desktop will have the same.

Bug ref:
https://github.com/readthedocs/sphinx_rtd_theme/issues/1452

Signed-off-by: Bill Mills <bill.mills@linaro.org>