github actions: run submodule sync only once per week
There are a lot of submodule updates happening.
To decrease maintenance commits run this job only once a week.
This still allows us to keep roughly up to date w/o a ton of extra commits.

If a submodule change is important to publish immediately the github action
can be run from the repo's web UI by any repo owner.

Signed-off-by: Bill Mills <bill.mills@linaro.org>