Doxygen: check in stock Doxygen layout file
This layout file is the built-in default for doxygen 1.9.1
It can be generated via the command:
$ doxygen -l

Check it in as is and then configure to use it.
There should be no difference in the output yet
(assuming that you are using doxygen 1.9.1 or similar).

Don't blame me for the trailing white space.
This is how stock doxygen had it.

Signed-off-by: Bill Mills <bill.mills@linaro.org>