overview: add AMP overview and introduce OpenAMP fundamentals
Provide a succinct introduction to AMP with comparison to SMP to lead into
a new section detailing the 'Fundamental' parts that OpenAMP is addressing
with the intent of leading into which parts are implemented by what
part of OpenAMP architecture which is the next section.
Parts taken from 'System Wide Considerations'.