contributing: squash contributing page and reference project governance
Most of this page is in governance, so remove most and leave only more
specific links to other sections providing more specific details on how to
contribute.

Refer issue 38: https://github.com/OpenAMP/openamp-docs/issues/38
Addressing review comments:
* Avoid duplication with https://www.openampproject.org/governance/ in
this pages. to determine if a part of the gouvernance should be moved
ghere
* no roadmap defined, to remove chapter
* no platform maintainer defined, to remove chapter and/or rewrite
according to gouvernance page