demos: correct multi services control flow and add sections
address review from Arnaud
"The diagram is not accurate for the ./rpmsg_export_dev usage. Between
rpmsg_ctrl and app_rpmsg_tty, there is a Name Service (NS) RPmsg from the
main processor to the remote processor(inverted on diagram).
In other words, rpmsg_export_dev allows creating an rpmsg TTY (or raw)
device instance on the main processor. In such a case, the main processor
sends an NS rpmsg message to the remote processor to inform it of the
creation.

Another suggestion: To clarify the diagram, it would be helpful to add
boxes or notes to separate and explain the subsequences. This would help
readers understand the relation between the demos and the diagram."