cleanup: remove future and current folders
These two folders look to have been a combination of work in progress
and todo items.
The pages do not seem accessible from the main content as neither index.rst
files are in a toc, but content within them is searchable which
results in the reader being able finding these pages. Remove to avoid
confusion for near future and re-introduce content as needed.