docs: create data structures file in readthedocs and use doxygen references
the data_structures file is a markdown (md) file in open-amp repository
which uses code within the file to demonstrate datastructures.
Review comments were to use doxygen linking, so use breathe and doxylink
in a new file within readthedocs to achieve the same reference but
through linking which will maintain itself as long as doxygen comments
are maintained in open-amp repo.