images: copy design document images from open-amp repository
Move all design document images from open-amp repository as it was
decided with Bill and Arnaud to move all the design documentation from the
https://github.com/OpenAMP/open-amp repository doc folder to the
https://github.com/OpenAMP/openamp-docs repository docs folder
The main reason being that the breathe and doxylink integration is already
in openamp-docs, so can be used for the design doc.
Also adjust the image links in the documentation.
Once this commit is merged, the documentation in open-amp can be deleted.

Signed-off-by: Sipke Vriend <sipke@direktembedded.com>